requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
tensorflow
pillow
numpy
python-multipart
aiofiles